Allow intermediate points to be converted to abeam points
NVU and NAS are easier if legs are longer, and at the time it was customary
to have longer legs. The NVU flight plan would only include points where you actually change
your navigation track by a meaningful amount. Let's add the same feature!

Abeam points will be displayed in the table with a small indentation